+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 3.0.0 (2020-01-21)
+
+-   Added eslint comments and import plugins
+-   Backported more settings from xo
+-   Bumped minimal Node version to 10
+
 ## 2.0.0 (2020-01-19)
 
 -   Added ava, unicorn and mocha plugins